+++ src/components/pages/apps/genes/gsea/ext-gsea/svg/es-svg.tsx
@@ -1,4 +1,4 @@
-import { useMemo, type ReactNode } from 'react'
+import { type ReactNode } from 'react'
 
 import { axisDomainToRangeFunc } from '@/components/plot/axes/axis'
 import { AxisBottomSvg, AxisLeftSvg } from '@/components/plot/axes/svg-axis'
@@ -6,12 +6,12 @@ import type { IExtGseaResult, IGseaResult } from '@/lib/gsea/ext-gsea'
 
 import { useAxis } from '@/components/plot/axes/axes-store'
 import { SvgG } from '@/components/plot/svg-g'
-import { SvgPolyLine } from '@/components/plot/svg-poly-line'
 import { SvgText } from '@/components/plot/svg-text'
-import { IPos } from '@/interfaces/pos'
 import { COLOR_BLACK } from '@/lib/color/color'
 import { type IGeneSet, type IRankedGene } from '@/lib/gsea/geneset'
+import { range } from '@/lib/math/range'
 import { where } from '@/lib/math/where'
+import { EsCurveSvg, EsLeadingEdgeSvg } from '../../gsea-plot/svg/es-svg'
 import { IExtGseaPlotResult, useExtGseaContext } from '../ext-gsea-provider'
 import { IExtGseaSettings } from '../ext-gsea-settings'
 
@@ -40,103 +40,129 @@ export function ExtGseaEsCurveSvg({
     axisId: 'y',
   })
 
+  if (gsea.leadingEdge.length === 0) {
+    return null
+  }
+
   const displayProps: IExtGseaSettings = plot.props
 
-  const curveSvg = useMemo(() => {
-    // size of plot with padding
+  // size of plot with padding
 
-    let hitPoints: IPos[] = gsea.esHits.map((g) => ({ x: g.rank, y: g.score }))
+  const ix = range(0, gsea.esAll.length, displayProps.es.step)
 
-    //range(y.length)
+  let subSampledRankedGenes = ix.map((i) => gsea.esAll[i]!)
 
-    // subsample so we don't draw every point
-    //const ix = range(0, x.length, displayProps.es.step)
-    //
-    //const x1 = ix.map((i) => x[i]!)
+  // let points: IPos[] = subSampledRankedGenes.map((g) => ({
+  //   x: g.rank,
+  //   y: g.score,
+  // }))
 
-    // we must end at the last point so zero the ends and fix
-    // fix x
+  //let hitPoints: IPos[] = gsea.esHits.map((g) => ({ x: g.rank, y: g.score }))
 
-    if (hitPoints[0].x !== 0) {
-      hitPoints = [{ x: 0, y: 0 }, ...hitPoints]
-    }
+  // subsample so we don't draw every point
+  //const ix = range(0, x.length, displayProps.es.step)
+  //
+  //const x1 = ix.map((i) => x[i]!)
 
-    if (hitPoints[hitPoints.length - 1].x !== result.rankedGenes.length - 1) {
-      hitPoints = [...hitPoints, { x: result.rankedGenes.length - 1, y: 0 }]
-    }
+  // we must end at the last point so zero the ends and fix
+  // fix x
 
-    //let y1 = ix.map((i) => y[i]!)
-    //y1[0] = 0
-    //y1[y1.length - 1] = 0
+  if (subSampledRankedGenes[0].rank !== 0) {
+    subSampledRankedGenes = [
+      { rank: 0, name: '', score: 0 },
+      ...subSampledRankedGenes,
+    ]
+  }
 
-    let leadingEdgeIdx =
-      gsea.es >= 0
-        ? gsea.leadingEdge[gsea.leadingEdge.length - 1].rank
-        : gsea.leadingEdge[0].rank
+  if (
+    subSampledRankedGenes[subSampledRankedGenes.length - 1].rank !==
+    result.rankedGenes.length - 1
+  ) {
+    subSampledRankedGenes = [
+      ...subSampledRankedGenes,
+      { rank: result.rankedGenes.length - 1, name: '', score: 0 },
+    ]
+  }
 
-    // now we want the ix that are within the leading edge
-    let leadingIdx =
-      gsea.es >= 0
-        ? where(hitPoints, (p) => p.x <= leadingEdgeIdx)
-        : where(hitPoints, (p) => p.x >= leadingEdgeIdx)
+  let leadingEdgeIdx =
+    gsea.es >= 0
+      ? gsea.leadingEdge[gsea.leadingEdge.length - 1].rank
+      : gsea.leadingEdge[0].rank
 
-    let lead = leadingIdx.map((i) => hitPoints[i])
+  // now we want the ix that are within the leading edge
+  let leadingIdx =
+    gsea.es >= 0
+      ? where(subSampledRankedGenes, (p) => p.rank <= leadingEdgeIdx)
+      : where(subSampledRankedGenes, (p) => p.rank >= leadingEdgeIdx)
 
-    // fix ends
+  const leadingEs = subSampledRankedGenes.filter((g) =>
+    gsea.es >= 0 ? g.rank <= leadingEdgeIdx : g.rank >= leadingEdgeIdx
+  )
 
-    lead =
-      gsea.es >= 0
-        ? [...lead, { ...lead[lead.length - 1]!, y: 0 }]
-        : [{ ...lead[0]!, y: 0 }, ...lead]
+  let leadingRankedGenes = leadingIdx.map((i) => subSampledRankedGenes[i])
 
-    let leadingEdge1Svg: ReactNode | undefined = undefined
+  // fix ends
 
-    const xaf = axisDomainToRangeFunc(xax)
-    const yaf = axisDomainToRangeFunc(yaxEs)
+  leadingRankedGenes =
+    gsea.es >= 0
+      ? [
+          ...leadingRankedGenes,
+          { ...leadingRankedGenes[leadingRankedGenes.length - 1]!, score: 0 },
+        ]
+      : [{ ...leadingRankedGenes[0]!, score: 0 }, ...leadingRankedGenes]
 
-    if (displayProps.es[gsMode].leadingEdge.show) {
-      const points = lead.map((p) => ({ x: xaf(p.x), y: yaf(p.y) }))
+  let leadingEdge1Svg: ReactNode | undefined = undefined
 
-      // const xs = axisDomainToRange(xax, xlead)
-      // const ys = axisDomainToRange(yaxEs, ylead)
+  const xaf = axisDomainToRangeFunc(xax)
+  const yaf = axisDomainToRangeFunc(yaxEs)
 
-      const pointsStr = points.map((p) => `${p.x},${p.y}`).join(' ')
-
-      leadingEdge1Svg = (
-        <SvgPolyLine
-          points={pointsStr}
-          fill={gs.color ?? displayProps.es[gsMode].leadingEdge.value}
-          fillOpacity={displayProps.es[gsMode].leadingEdge.opacity}
-          stroke="none"
-        />
-      )
-    }
-
-    let line1Svg: ReactNode | undefined = undefined
-
-    if (displayProps.es[gsMode].curve.show) {
-      const points = hitPoints.map((p) => ({ x: xaf(p.x), y: yaf(p.y) }))
-      const pointsStr = points.map((p) => `${p.x},${p.y}`).join(' ')
-
-      line1Svg = (
-        <SvgPolyLine
-          points={pointsStr}
-          stroke={gs.color ?? displayProps.es[gsMode].curve.value}
-          s={displayProps.es[gsMode].curve}
-          fill="none"
-        />
-      )
-    }
-
-    return (
-      <>
-        {leadingEdge1Svg && leadingEdge1Svg}
-        {line1Svg && line1Svg}
-      </>
+  if (displayProps.es[gsMode].leadingEdge.show) {
+    leadingEdge1Svg = (
+      <EsLeadingEdgeSvg
+        leadingEdge={leadingEs}
+        xaf={xaf}
+        yaf={yaf}
+        fill={gs.color ?? displayProps.es[gsMode].leadingEdge.value}
+        fillOpacity={displayProps.es[gsMode].leadingEdge.opacity}
+      />
     )
-  }, [xax, yaxEs, displayProps, result, gs, gsea, gsMode])
+  }
 
-  return curveSvg
+  let line1Svg: ReactNode | undefined = undefined
+
+  if (displayProps.es[gsMode].curve.show) {
+    const points = subSampledRankedGenes.map((p) => ({
+      x: xaf(p.rank),
+      y: yaf(p.score),
+    }))
+    //const pointsStr = points.map((p) => `${p.x},${p.y}`).join(' ')
+
+    line1Svg = (
+      <EsCurveSvg
+        es={gsea.esHits}
+        points={points}
+        xax={xax}
+        yax={yaxEs}
+        stroke={gs.color ?? displayProps.es[gsMode].curve.value}
+      />
+    )
+
+    // line1Svg = (
+    //   <SvgPolyLine
+    //     points={pointsStr}
+    //     stroke={gs.color ?? displayProps.es[gsMode].curve.value}
+    //     s={displayProps.es[gsMode].curve}
+    //     fill="none"
+    //   />
+    // )
+  }
+
+  return (
+    <>
+      {leadingEdge1Svg && leadingEdge1Svg}
+      {line1Svg && line1Svg}
+    </>
+  )
 }
 
 export function ExtGseaEsSvgPlot({ result }: { result: IExtGseaPlotResult }) {
@@ -156,116 +182,103 @@ export function ExtGseaEsSvgPlot({ result }: { result: IExtGseaPlotResult }) {
 
   const displayProps: IExtGseaSettings = plot.props
 
-  const esSvg = useMemo(() => {
-    const rankedGenes: IRankedGene[] = result.rankedGenes
-    const gs1: IGeneSet = result.gs1
-    const gs2: IGeneSet = result.gs2
+  const rankedGenes: IRankedGene[] = result.rankedGenes
+  const gs1: IGeneSet = result.gs1
+  const gs2: IGeneSet = result.gs2
 
-    const extGsea: IExtGseaResult = result.extGsea
-    const gsea1: IGseaResult = result.gsea1
-    const gsea2: IGseaResult = result.gsea2
+  const extGsea: IExtGseaResult = result.extGsea
+  const gsea1: IGseaResult = result.gsea1
+  const gsea2: IGseaResult = result.gsea2
 
-    const yaf = axisDomainToRangeFunc(yaxEs)
+  const yaf = axisDomainToRangeFunc(yaxEs)
 
-    const esSvg = (
+  if (!gsea1 || !gsea2) {
+    return null
+  }
+
+  return (
+    <SvgG
+      pos={{
+        x: 0,
+        y: 0,
+      }}
+    >
+      <ExtGseaEsCurveSvg result={result} gs={gs1} gsea={gsea1} gsMode="gs1" />
+
+      <ExtGseaEsCurveSvg result={result} gs={gs2} gsea={gsea2} gsMode="gs2" />
+
+      <AxisLeftSvg ax={yaxEs} />
       <SvgG
         pos={{
           x: 0,
-          y: 0,
+          y: yaf(0),
         }}
       >
-        <ExtGseaEsCurveSvg result={result} gs={gs1} gsea={gsea1} gsMode="gs1" />
-
-        <ExtGseaEsCurveSvg result={result} gs={gs2} gsea={gsea2} gsMode="gs2" />
-
-        <AxisLeftSvg ax={yaxEs} />
+        <AxisBottomSvg ax={xax} showTicks={displayProps.es.axes.x.showTicks} />
         <SvgG
           pos={{
-            x: 0,
-            y: yaf(0),
+            x: displayProps.axes.x.length + displayProps.plot!.gap.x / 2,
+            y: 0,
           }}
         >
-          <AxisBottomSvg
-            ax={xax}
-            showTicks={displayProps.es.axes.x.showTicks}
-          />
-          <SvgG
-            pos={{
-              x: displayProps.axes.x.length + displayProps.plot!.gap.x / 2,
-              y: 0,
-            }}
-          >
-            <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
-              {rankedGenes.length.toLocaleString()}
-            </SvgText>
-          </SvgG>
+          <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
+            {rankedGenes.length.toLocaleString()}
+          </SvgText>
         </SvgG>
-
-        <SvgG
-          pos={{
-            x: 0,
-            y: displayProps.es.axes.y.length + displayProps.plot!.gap.y / 2,
-          }}
-        >
-          <SvgG>
-            <SvgText
-              fill={COLOR_BLACK}
-              font={displayProps.axes.x.font}
-              //fontSize="x-small"
-              //textAnchor="middle"
-              //fontWeight="bold"
-            >
-              {gs1.name}
-            </SvgText>
-          </SvgG>
-
-          <SvgG
-            pos={{
-              x: displayProps.axes.x.length,
-              y: 0,
-            }}
-          >
-            <SvgText
-              fill={COLOR_BLACK}
-              font={displayProps.axes.x.font}
-              //fontSize="x-small"
-              textAnchor="end"
-              //fontWeight="bold"
-            >
-              {gs2.name}
-            </SvgText>
-          </SvgG>
-        </SvgG>
-
-        {displayProps.es.stats.show && (
-          <SvgG
-            id="stats"
-            pos={{
-              x: displayProps.axes.x.length,
-              y: 0,
-            }}
-          >
-            <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
-              NES: {extGsea.nes.toFixed(2)}
-            </SvgText>
-
-            <SvgG
-              pos={{
-                x: 0,
-                y: 20,
-              }}
-            >
-              <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
-                P-value: {extGsea.pvalue.toFixed(3)}
-              </SvgText>
-            </SvgG>
-          </SvgG>
-        )}
       </SvgG>
-    )
 
-    return esSvg
-  }, [xax, yaxEs, result, displayProps])
+      <SvgG
+        pos={{
+          x: 0,
+          y: displayProps.es.axes.y.length + displayProps.plot!.gap.y / 2,
+        }}
+      >
+        <SvgG>
+          <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
+            {gs1.name}
+          </SvgText>
+        </SvgG>
 
-  return esSvg
+        <SvgG
+          pos={{
+            x: displayProps.axes.x.length,
+            y: 0,
+          }}
+        >
+          <SvgText
+            fill={COLOR_BLACK}
+            font={displayProps.axes.x.font}
+            textAnchor="end"
+          >
+            {gs2.name}
+          </SvgText>
+        </SvgG>
+      </SvgG>
+
+      {displayProps.es.stats.show && (
+        <SvgG
+          id="stats"
+          pos={{
+            x: displayProps.axes.x.length,
+            y: 0,
+          }}
+        >
+          <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
+            NES: {extGsea.nes.toFixed(2)}
+          </SvgText>
+
+          <SvgG
+            pos={{
+              x: 0,
+              y: 20,
+            }}
+          >
+            <SvgText fill={COLOR_BLACK} font={displayProps.axes.x.font}>
+              P-value: {extGsea.pvalue.toFixed(3)}
+            </SvgText>
+          </SvgG>
+        </SvgG>
+      )}
+    </SvgG>
+  )
 }