+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- ext gsea shares more code with original gsea plot. Ext gsea now plots the full es for the curve to make line smoother.
 - Smaller header links grid.
 - crosshair supports content. crosshairs now in fixed positions for gsea.
 - More removal of duplicate code in ext gsea and optimizing dependencies.